@@ -1,52 +1,54 @@
 package com.java.retrospective.services.impl;
 
-import com.java.retrospective.dao.RetrospectiveDao;
-import com.java.retrospective.dao.SwimlaneDao;
-import com.java.retrospective.dto.swimlane.SwimlaneDto;
-import com.java.retrospective.dto.user.UserDto;
-import com.java.retrospective.entity.RetrospectiveEntity;
-import com.java.retrospective.entity.SwimlaneEntity;
-import com.java.retrospective.entity.UserEntity;
-import com.java.retrospective.services.SwimlaneService;
-import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RestController;
-
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import com.java.retrospective.dao.RetrospectiveDao;
+import com.java.retrospective.dao.SwimlaneDao;
+import com.java.retrospective.dto.retrospeective.RetrospectiveOutDto;
+import com.java.retrospective.dto.swimlane.SwimlaneInDto;
+import com.java.retrospective.dto.swimlane.SwimlaneOutDto;
+import com.java.retrospective.entity.RetrospectiveEntity;
+import com.java.retrospective.entity.SwimlaneEntity;
+import com.java.retrospective.mappers.SwimlaneMapper;
+import com.java.retrospective.services.RetrospectiveService;
+import com.java.retrospective.services.SwimlaneService;
+
+import com.java.retrospective.validator.RetrospectiveValidator;
+import com.java.retrospective.validator.SwimlaneValidator;
+import org.mapstruct.Mapper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import lombok.AllArgsConstructor;
 
 @Service
 @AllArgsConstructor
 public class SwimlaneServiceImpl implements SwimlaneService {
 
-    @Autowired
-    private SwimlaneDao swimlaneDao;
-    @Autowired
-    private RetrospectiveDao retrospectiveDao;
+    private final SwimlaneDao swimlaneDao;
+    private final SwimlaneMapper swimlaneMapper;
+    private final RetrospectiveService retrospectiveService;
+    private final SwimlaneValidator swimlaneValidator;
 
     @Override
-    public List<SwimlaneDto> getAllSwimlanes(Integer retro_id) {
-//        List<SwimlaneEntity> swimlaneEntities = swimlaneDao.findAll().stream().filter(swimlaneEntity -> swimlaneEntity.getRetrospective().getId().equals(retro_id)).collect(Collectors.toList());
-//        return swimlaneEntities.stream().map(this::convertDataIntoDto).collect(Collectors.toList());
+    public List<SwimlaneOutDto> getAllSwimlanes(Integer retro_id) {
         List<SwimlaneEntity> swimlaneEntities = swimlaneDao.findByRetrospectiveId(retro_id);
-        return  swimlaneEntities.stream().map(this::convertDataIntoDto).collect(Collectors.toList());
+        return  swimlaneEntities==null? null:swimlaneEntities.stream().map(swimlaneMapper::mapEntityToOutDto).collect(Collectors.toList());
     }
 
     @Override
-    public SwimlaneDto addSwimlane(Integer retro_id, String swimlaneTitle) {
-        Optional<RetrospectiveEntity> optionalRetrospectiveEntity= retrospectiveDao.findById(retro_id);
-        if(optionalRetrospectiveEntity.isPresent()){
-            RetrospectiveEntity retrospectiveEntity = optionalRetrospectiveEntity.get();
-            SwimlaneEntity swimlaneEntity = new SwimlaneEntity();
-            swimlaneEntity.setRetrospective(retrospectiveEntity);
-            swimlaneEntity.setTitle(swimlaneTitle);
-            return convertDataIntoDto(swimlaneDao.save(swimlaneEntity));
-        }else{
-            return null;
+    public SwimlaneOutDto addSwimlane(Integer retro_id, SwimlaneInDto swimlane) {
+        if(swimlaneValidator.validateForCreate(swimlane)) {
+            RetrospectiveEntity retrospectiveEntity = retrospectiveService.getRetrospectiveEntity(retro_id);
+            if (retrospectiveEntity != null) {
+                SwimlaneEntity swimlaneEntity = swimlaneMapper.mapInDtoToEntity(swimlane);
+                swimlaneEntity.setRetrospective(retrospectiveEntity);
+                return swimlaneMapper.mapEntityToOutDto(swimlaneDao.save(swimlaneEntity));
+            }
         }
-
+        return null;
     }
 
     @Override
@@ -54,10 +56,10 @@ public class SwimlaneServiceImpl implements SwimlaneService {
         swimlaneDao.deleteById(id);
         return "Swimlane deleted!";
     }
-    private SwimlaneDto convertDataIntoDto(SwimlaneEntity swimlane){
-        SwimlaneDto swimlaneDto= new SwimlaneDto();
-        swimlaneDto.setTitle(swimlane.getTitle());
-        swimlaneDto.setRetrospective(swimlane.getRetrospective());
-        return swimlaneDto;
+
+    @Override
+    public SwimlaneEntity getSwimlaneById(Integer id) {
+        return swimlaneDao.findById(id).orElse(null);
     }
+
 }
