@@ -1,31 +1,36 @@
 package com.java.retrospective.controller;
 
+import java.util.List;
+
 import com.java.retrospective.dto.user.UserDto;
 import com.java.retrospective.services.UserService;
-import com.java.retrospective.services.mappers.UserMappingService;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
-import java.util.List;
 
 @RestController
 @RequestMapping(value = "/users")
 @AllArgsConstructor
 public class UserController {
-    @Autowired
-    private UserService userService;
-    @Autowired
-    private UserMappingService userServicerMappingService;
+    private final UserService userService;
+
     @PostMapping
     public String addUsers(@RequestBody UserDto user){
-        userService.addUser(user);
-        return "User registration successful!";
+        if(userService.addUser(user)!=null)
+            return "User registration successful!";
+        return "Incorrect inputs. Please check the username and password.";
     }
 
     @GetMapping
     public List<UserDto> getAllUsers(){
-        List<UserDto> userDtos=userService.getAllUsers();
+        List<UserDto> userDtos = userService.getAllUsers();
         return userDtos;
     }
+
 
 }
