@@ -1,18 +1,25 @@
 package com.java.retrospective.entity;
 
-import lombok.*;
-import org.springframework.web.bind.annotation.RequestMapping;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
 
-import javax.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
 
 @Entity
 @Table(name="message")
 @NoArgsConstructor
 @AllArgsConstructor
-@Getter
-@Setter
-@EqualsAndHashCode
-@ToString
+@Data
 public class MessageEntity {
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
@@ -22,51 +29,11 @@ public class MessageEntity {
     @NonNull private String message;
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "ref_user", referencedColumnName = "id")
-    @NonNull private UserEntity user;
+    @NonNull private UserEntity user=null;
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "ref_retrospective", referencedColumnName = "id")
-    @NonNull private RetrospectiveEntity retrospective;
+    @NonNull private RetrospectiveEntity retrospective=null;
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "ref_swimlane", referencedColumnName = "id")
-    @NonNull private SwimlaneEntity swimlane;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public UserEntity getUser() {
-        return user;
-    }
-
-    public void setUser(UserEntity user) {
-        this.user = user;
-    }
-
-    public RetrospectiveEntity getRetrospective() {
-        return retrospective;
-    }
-
-    public void setRetrospective(RetrospectiveEntity retrospective) {
-        this.retrospective = retrospective;
-    }
-
-    public SwimlaneEntity getSwimlane() {
-        return swimlane;
-    }
-
-    public void setSwimlane(SwimlaneEntity swimlane) {
-        this.swimlane = swimlane;
-    }
+    @NonNull private SwimlaneEntity swimlane=null;
 }
