@@ -1,17 +1,27 @@
 package com.java.retrospective.entity;
 
-import lombok.*;
-
-import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
 
 @Entity
 @Table(name="retrospective")
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-@ToString
 public class RetrospectiveEntity {
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
@@ -22,37 +32,6 @@ public class RetrospectiveEntity {
     @Column
     private String description;
     @OneToMany(mappedBy = "retrospective")
+    @JsonManagedReference
     @NonNull private List<SwimlaneEntity> swimlanes= new ArrayList<>();
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public List<SwimlaneEntity> getSwimlanes() {
-        return swimlanes;
-    }
-
-    public void setSwimlanes(List<SwimlaneEntity> swimlanes) {
-        this.swimlanes = swimlanes;
-    }
 }
