@@ -1,16 +1,22 @@
 package com.java.retrospective.entity;
 
-import lombok.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
 
-import javax.persistence.*;
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import lombok.*;
 
 @Entity
 @Table(name="swimlane")
 @NoArgsConstructor
 @AllArgsConstructor
-@Getter
-@Setter
-@ToString
+@Data
 public class SwimlaneEntity {
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
@@ -21,29 +27,7 @@ public class SwimlaneEntity {
 
     @ManyToOne
     @JoinColumn(name = "ref_retrospective", referencedColumnName = "id")
-    private RetrospectiveEntity retrospective;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public RetrospectiveEntity getRetrospective() {
-        return retrospective;
-    }
-
-    public void setRetrospective(RetrospectiveEntity retrospective) {
-        this.retrospective = retrospective;
-    }
+    @JsonBackReference
+    @ToString.Exclude
+    private RetrospectiveEntity retrospective=null;
 }
