@@ -1,6 +1,7 @@
 package com.java.retrospective.services;
 
-import com.java.retrospective.dto.message.MessageDto;
+import com.java.retrospective.dto.message.MessageInDto;
+import com.java.retrospective.dto.message.MessageOutDto;
 import com.java.retrospective.entity.MessageEntity;
 import org.springframework.stereotype.Service;
 
@@ -8,8 +9,8 @@ import java.util.List;
 
 public interface MessageService {
 
-    List<MessageDto> getAllMessages(Integer retro_id);
-    MessageDto addMessage(Integer retro_id, MessageEntity messageEntity);
+    List<MessageOutDto> getAllMessages(Integer retro_id);
+    MessageOutDto addMessage(Integer retro_id, MessageInDto messageEntity);
     String deleteMessage(Integer retro_id, Integer id);
-    MessageDto updateMessage(Integer retro_id, Integer id, String message);
+    MessageOutDto updateMessage(Integer retro_id, Integer id, String message);
 }
