@@ -1,14 +1,15 @@
 package com.java.retrospective.services;
 
-import com.java.retrospective.dto.swimlane.SwimlaneDto;
+import com.java.retrospective.dto.swimlane.SwimlaneInDto;
+import com.java.retrospective.dto.swimlane.SwimlaneOutDto;
 import com.java.retrospective.entity.SwimlaneEntity;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 public interface SwimlaneService {
-    List<SwimlaneDto> getAllSwimlanes(Integer retro_id);
-    SwimlaneDto addSwimlane(Integer retro_id, String swimlaneEntity);
+    List<SwimlaneOutDto> getAllSwimlanes(Integer retro_id);
+    SwimlaneOutDto addSwimlane(Integer retro_id, SwimlaneInDto swimlane);
     String deleteSwimlane(Integer retro_id, Integer id);
+    SwimlaneEntity getSwimlaneById(Integer id);
 
 }
